pt: dtype

+++ pytraj/api.py
@@ -568,8 +568,8 @@ class Trajectory(ActionTrajectory):
             frame.trans_rot_trans(v1, mat, v2)
             self._xyz[idx] = frame.xyz
 
-    def _allocate(self, n_frames, n_atoms):
-        self._xyz = np.empty((n_frames, n_atoms, 3))
+    def _allocate(self, n_frames, n_atoms, dtype='f8'):
+        self._xyz = np.empty((n_frames, n_atoms, 3), dtype=dtype)
 
     def strip_atoms(self, mask):
         atm = self.top.select(mask)
